Add console grep-like utility
This commits adds a klogg_grep console tool (#306). It uses klogg search
engine to find all matches for given pattern in a file.

It is intended to be used for acceptance testing. It can't replace
grep-like utilities because klogg_grep builds an index of given file
and only after that does pattern matching. Indexing beforehand is great
if one needs to scan for several patterns, but completely unnecessary
for single search in a file.