feature: allow files with more than 2147483647 lines (#339, #341)
This commit switches internal types for line numbers to `uint64_t`. This allows to handle files with more than 2147483647 lines.

For files with smaller number of lines there should be no visible changes in behavior. For larger files there is a problem with scrolling because Qt scrollbar is limited to `int` number of steps.